Add bit depth handling to audio track and related components
- Introduced a new `BitDepth` enum to represent audio sample formats and bit depths, enhancing the audio track structure.
- Updated multiple modules to include `bit_depth` as an optional field in audio track definitions, accommodating lossy codecs that do not provide this information.
- Modified audio processing functions to handle `f32` samples instead of `i16`, improving precision during audio operations.
- Enhanced tests across various modules to validate the integration of bit depth handling and ensure consistent behavior in audio processing.

These changes aim to improve audio fidelity and maintain clarity in the handling of bit depth during audio processing.

+++ crates/clip-sync-repair/src/application/patch_region.rs
@@ -50,7 +50,7 @@ pub(crate) struct SeamGateOutcome {
 
 pub(crate) struct SeamGateParams<'a> {
     pub a_pcm: &'a MultiChannelPcm,
-    pub b_samples: &'a [i16],
+    pub b_samples: &'a [f32],
     pub b_extract_start_secs: f64,
     pub refined_b_start_secs: f64,
     pub refined_b_end_secs: f64,
